pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pix2pix-gan"
version = "2.0.0"
description = "Image-to-Image Translation with Conditional GANs (Pix2Pix)"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Vishnu Nambiar", email = "vishnu@example.com"},
]
keywords = [
    "pix2pix",
    "gan",
    "generative-adversarial-network",
    "image-translation",
    "deep-learning",
    "pytorch",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: GPU :: NVIDIA CUDA",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Typing :: Typed",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "Pillow>=9.0.0",
    "scikit-image>=0.19.0",
    "opencv-python>=4.5.0",
    "matplotlib>=3.5.0",
    "tqdm>=4.62.0",
    "tensorboard>=2.8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "isort>=5.11",
    "flake8>=4.0",
    "pylint>=2.15",
    "mypy>=0.990",
    "pre-commit>=2.20",
]

evaluation = [
    "lpips>=0.1.4",
]

docs = [
    "sphinx>=5.0",
    "sphinx-rtd-theme>=1.0",
    "sphinx-autodoc-typehints>=1.19",
]

[project.urls]
Homepage = "https://github.com/VishnuNambiar0602/Image_to_image_translation"
Documentation = "https://image-to-image-translation.readthedocs.io"
Repository = "https://github.com/VishnuNambiar0602/Image_to_image_translation.git"
Issues = "https://github.com/VishnuNambiar0602/Image_to_image_translation/issues"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py10", "py11", "py12"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | datasets
)/
'''

[tool.isort]
profile = "black"
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100
skip = [".git", ".venv", "venv", "datasets"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "torchvision.*",
    "tensorboard.*",
    "lpips.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short --strict-markers --cov=src/pix2pix"
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "slow: slow tests",
]

[tool.coverage.run]
source = ["src/pix2pix"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
    "R0903",  # too-few-public-methods
]

[tool.pylint.format]
max-line-length = 100
